# web/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Evidence-first agricultural extension officer workbench for Benin.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /plots/{plotId}/candidates:
    get:
      operationId: getPlotFieldCandidates
      tags: [plots]
      summary: Satellite-detected fields near a plot, ranked, for an officer to confirm
      parameters:
        - { in: path, name: plotId, required: true, schema: { type: string } }
      responses:
        "200":
          description: Ranked candidates (suggestions only)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FieldCandidates" }
        "404": { description: Plot not found }
  /plots/{plotId}/match:
    post:
      operationId: matchPlotField
      tags: [plots]
      summary: An officer confirms that a detected field is this household's plot
      parameters:
        - { in: path, name: plotId, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MatchFieldInput" }
      responses:
        "200":
          description: Plot geometry updated, verification level officer_matched
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlotMatchResult" }
        "400": { description: Invalid officer name }
        "404": { description: Plot or candidate not found }
        "409": { description: Plot protected (surveyed/cadastral) or field already matched }
...
components:
  schemas:
...
    FieldCandidate:
      type: object
      required: [candidateId, source, areaHa, distanceM, score, reasons, alreadyMatched, geometry]
      properties:
        candidateId: { type: string }
        source: { type: string }
        areaHa: { type: number }
        distanceM: { type: integer }
        score: { type: number }
        reasons: { type: array, items: { type: string } }
        alreadyMatched: { type: boolean }
        geometry: { type: object, additionalProperties: true }
    FieldCandidates:
      type: object
      required: [plotId, verificationLevel, origin, candidates, radiusKm, note]
      properties:
        plotId: { type: string }
        verificationLevel: { type: string }
        declaredAreaHa: { type: ["number", "null"] }
        origin:
          type: object
          required: [lat, lon, hasPolygon]
          properties:
            lat: { type: number }
            lon: { type: number }
            hasPolygon: { type: boolean }
        candidates: { type: array, items: { $ref: "#/components/schemas/FieldCandidate" } }
        radiusKm: { type: number }
        note: { type: string }
    MatchFieldInput:
      type: object
      required: [candidateId, officerName]
      properties:
        candidateId: { type: string }
        officerName: { type: string, minLength: 2, maxLength: 120 }
    PlotMatchResult:
      type: object
      required: [plotId, candidateId, verificationLevel, previousVerification, weatherCellId, matchedBy]
      properties:
        plotId: { type: string }
        candidateId: { type: string }
        verificationLevel: { type: string }
        previousVerification: { type: string }
        weatherCellId: { type: string }
        matchedBy: { type: string }
...