# web/lib/api-zod/src/generated/types/fieldCandidate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Evidence-first agricultural extension officer workbench for Benin.
 * OpenAPI spec version: 0.1.0
 */
import type { FieldCandidateGeometry } from './fieldCandidateGeometry';

export interface FieldCandidate {
  candidateId: string;
  source: string;
  areaHa: number;
  distanceM: number;
  score: number;
  reasons: string[];
  alreadyMatched: boolean;
  geometry: FieldCandidateGeometry;
}