# web/lib/api-zod/src/generated/types/fieldCandidates.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Evidence-first agricultural extension officer workbench for Benin.
 * OpenAPI spec version: 0.1.0
 */
import type { FieldCandidate } from './fieldCandidate';
import type { FieldCandidatesOrigin } from './fieldCandidatesOrigin';

export interface FieldCandidates {
  plotId: string;
  verificationLevel: string;
  /** @nullable */
  declaredAreaHa?: number | null;
  origin: FieldCandidatesOrigin;
  candidates: FieldCandidate[];
  radiusKm: number;
  note: string;
}