# web/lib/api-zod/src/generated/types/matchFieldInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Evidence-first agricultural extension officer workbench for Benin.
 * OpenAPI spec version: 0.1.0
 */

export interface MatchFieldInput {
  candidateId: string;
  /**
     * @minLength 2
     * @maxLength 120
     */
  officerName: string;
}