# web/lib/api-zod/src/generated/types/plotMatchResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Evidence-first agricultural extension officer workbench for Benin.
 * OpenAPI spec version: 0.1.0
 */

export interface PlotMatchResult {
  plotId: string;
  candidateId: string;
  verificationLevel: string;
  previousVerification: string;
  weatherCellId: string;
  matchedBy: string;
}